Add firebase token to header
